requirements.txt
fastapi
uvicorn
python-multipart
pydantic
numpy  
Pillow
google-cloud-storage
tensorflow==2.16.1 
keras==3.3.3


